Move `mbr.h` to `image_aggregator`
Bug: b/435760503
