+++ base/cvd/cuttlefish/host/libs/config/data_image.cpp
@@ -38,9 +38,9 @@
 #include "cuttlefish/host/libs/config/config_utils.h"
 #include "cuttlefish/host/libs/config/cuttlefish_config.h"
 #include "cuttlefish/host/libs/config/esp.h"
-#include "cuttlefish/host/libs/config/mbr.h"
 #include "cuttlefish/host/libs/config/openwrt_args.h"
 #include "cuttlefish/host/libs/config/vmm_mode.h"
+#include "cuttlefish/host/libs/image_aggregator/mbr.h"
 
 // https://cs.android.com/android/platform/superproject/main/+/main:device/google/cuttlefish/Android.bp;l=127;drc=6f7d6a4db58efcc2ddd09eda07e009c6329414cd
 #define F2FS_BLOCKSIZE "4096"
